+++ db/migrate/20191001194145_create_invoices.rb
@@ -1,8 +1,8 @@
 class CreateInvoices < ActiveRecord::Migration[5.1]
   def change
     create_table :invoices do |t|
-      t.integer :customer_id
-      t.integer :merchant_id
+      t.bigint :customer_id
+      t.bigint :merchant_id
       t.string :status
       t.string :created_at
       t.string :updated_at
