+++ db/migrate/20191001201636_create_invoice_items.rb
@@ -1,8 +1,8 @@
 class CreateInvoiceItems < ActiveRecord::Migration[5.1]
   def change
     create_table :invoice_items do |t|
-      t.integer :item_id
-      t.integer :invoice_id
+      t.bigint :item_id
+      t.bigint :invoice_id
       t.integer :quantity
       t.integer :unit_price
       t.string :created_at
